feat(transform): CHAT → TalkBank XML emitter + ancillary cleanup
Adds a library-only Rust XML emitter at `crates/talkbank-transform/src/xml/`
that produces TalkBank-XML matching the output of the legacy Java
Chatter tool. All 39 reference-XML goldens in `corpus/reference-xml/`
pass byte-equivalent structural comparison via the new
`xml_golden_parity` harness in `crates/talkbank-parser-tests/tests/
xml_golden.rs`.

## Emitter scope and layout

Legacy/rare-use facility: TalkBank no longer publishes XML for
download, and every external XML consumer has either migrated to
CHAT (Phon, TalkBankDB) or gone dormant (NLTK CHILDESCorpusReader,
langcog/childes-db). The emitter exists to support the rare legacy
case; new integrations should consume CHAT directly.

- Emission only; XML ingest is explicitly out of scope.
- Phonetic / syllabification tiers (%pho, %mod, %phosyl, %modsyl,
  %phoaln) are permanently unsupported via
  `XmlWriteError::PhoneticTierUnsupported` — porting the phon-ipa
  IPA parser into Rust solely to round-trip them is not justified
  when no downstream consumer reads the rich
  `<pg>/<pw>/<ph>/<cmph>/<ss>` shape.

Module layout (single `XmlEmitter` state holder; each submodule
contributes its own `impl XmlEmitter { … }` block):

- `writer.rs` — `write_chat_xml` entry + `escape_text` (partial
  XML escape, apostrophes pass through literally).
- `root.rs` — document / participants / body / utterance
  orchestration; metadata helpers (corpus lookup, date/age/sex
  formatting, `@Options` flags, `@Types` projection onto root,
  per-speaker `@Birthplace` / `@L1` extras).
- `word.rs` — `<w>`, `<t>`, `<tagMarker>`, `<s>`, `<pause>`,
  `<linker>`, `<overlap-point>`, `<g>` wrappers, scoped
  annotations (`<error/>`, `<ga type="explanation">`,
  `<k type="stressing"/>`), word-internal markers.
- `mor.rs` — `<mor>` / `<mw>` / `<gra>` subtree, `<mor-post>`
  for post-clitic chains, `UtteranceTiers` aggregator.
- `wor.rs` — `<media>`, `<wor>`, `<internal-media>`;
  `format_seconds` ms → seconds.
- `deptier.rs` — text dependent tiers → `<a type="…">`.

## Two-cursor invariant

Walking the main tier requires two independent cursors into the
dependent tiers — a single cursor silently drifts on any utterance
containing a clitic chain:

- `mor_cursor` indexes `mor_tier.items`; advances by **1** per
  alignable word (clitic chain is one Mor item).
- `gra_chunk` is the 1-based `%gra` index; advances by
  **`1 + post_clitics.len()`** per Mor (each post-clitic
  contributes its own `<gra>` edge).

Cursor gating also honours the model's canonical
`counts_for_tier(word, TierDomain::Mor)` predicate so nonwords,
fillers, fragments, and untranscribed placeholders don't consume
chunks.

## Rust-vs-Java bridges (legitimate divergences)

Four places where the Rust model and Java Chatter semantically
disagree — all confirmed as legitimate format-evolution divergences,
not regressions. The emitter bridges at the output boundary:

1. CA contour terminators (⇗ ↗ → ↘ ⇘) emit `<s>` sibling +
   `<t type="missing CA terminator"/>` pair.
2. `CAOmission` (`(parens)`) renders as `<shortening>` wrapper
   rather than `type="omission"` on `<w>`.
3. Leading overlap-point markers (`⌈` / `⌊` at word start) are
   hoisted out as `<overlap-point/>` siblings of `<w>`; internal
   / trailing ones stay inside.
4. `xxx` / `yyy` / `www` case-sensitivity — the model's
   `word.untranscribed()` helper is case-insensitive as a
   Stanza/MOR correctness fix; the emitter uses a case-sensitive
   gate at the XML boundary since the schema attribute only
   attaches to lowercase placeholders.

## Supporting docs, specs, and goldens

- `book/src/architecture/xml-emitter.md` — developer-facing
  architecture doc (module boundaries, cursor invariants, bridge
  points, testing surface).
- `docs/reference-xml-coverage-gaps.md` — 65 of 98 reference
  `.cha` files have no paired XML golden; every gap is a Java
  Chatter rejection driven by CHAT format evolution, not a Rust
  bug. Methodology + categorised error inventory.
- `docs/rust-vs-java-chatter-regressions.md` — audit of the 245
  Java diagnostic lines; surfaces one candidate regression
  (`@Media` without bullets).
- `spec/errors/E544_media_linkage_without_timing.md` — proposed
  validator for that one candidate regression. Approved in
  principle; `Status: not_implemented`.
- `corpus/reference-xml/**/*.xml` — 33 Java Chatter goldens
  committed alongside eight `00errors.cex` diagnostic logs.
  `corpus/reference-xml/content/deprecated-skip-bullet.xml` is
  intentionally absent — the `skip="true"` feature was retired
  from CHAT on 2026-03-31 and the Rust model no longer preserves
  it. `corpus/reference-xml/README.md` documents the intentional
  omission policy.
- `crates/talkbank-parser-tests/tests/xml_golden.rs` +
  `xml_support/mod.rs` — structural-comparison harness
  (whitespace + attribute-order insensitive).

## Cleanup included in this commit

- Workspace-wide `cargo fmt` + `cargo clippy --all-targets` → zero
  warnings across all crates.
- Renamed `alignment::traits::positional_align` return tuple to a
  named `PositionalAlignment<T>` type alias (fixes clippy complex-type
  warning that had been outstanding).
- Idiom fixes in `talkbank-model`, `talkbank-parser`,
  `talkbank-re2c-parser`, `talkbank-lsp`, `talkbank-clan`, plus
  the top-level `talkbank-tools` bin — all mechanical clippy-fix
  applications (if-in-match guards, `sort_by_key` + `Reverse`,
  removed unnecessary `as usize` cast, removed duplicated
  `#[allow(dead_code)]`, etc.).
- `generated_traversal.rs` now allows `clippy::collapsible_if`
  (609 warnings the tree-sitter-grammar-utils generator produces;
  suppressed at file level until the upstream generator emits
  let-chains).
- Two unrelated in-progress changes in `talkbank-model/mor/analysis/`
  (new `clan_ud_mapping.rs` + a `Terminator::try_from_chat_str`
  helper) are folded into this commit.

+++ crates/talkbank-parser/src/parser/tree_parsing/header/id/parse.rs
@@ -45,7 +45,7 @@ pub fn parse_id_header(node: Node, source: &str, errors: &impl ErrorSink) -> Hea
     };
 
     // Extract fields by position from id_contents
-    let child_count = id_contents.child_count() as usize;
+    let child_count = id_contents.child_count();
     let mut idx = 0usize;
 
     let language = parse_required_text_field(
